+++ src/handlers/parsers.js
@@ -94,7 +94,7 @@ export const defaultGasPriceFormat = (option, timeAmount, valueAmount, valueDisp
  * @param {Object} prices
  * @param {Number} gasLimit
  */
-export const parseGasPrices = (data, prices, gasLimit, short) => {
+export const parseGasPrices = (data, prices, gasLimit, nativeCurrency, short) => {
   const gasPrices = {
     slow: null,
     average: null,
@@ -135,33 +135,36 @@ export const parseGasPrices = (data, prices, gasLimit, short) => {
       short
     );
   }
-  return parseGasPricesTxFee(gasPrices, prices, gasLimit);
+  return parseGasPricesTxFee(gasPrices, prices, gasLimit, nativeCurrency);
 };
 
-export const convertGasPricesToNative = (prices, gasPrices) => {
+export const convertGasPricesToNative = (prices, gasPrices, nativeCurrency) => {
   const nativeGases = { ...gasPrices };
-  if (prices && prices.selected) {
-    gasPrices.fast.txFee.native = getNativeGasPrice(prices, gasPrices.fast.txFee.value.amount);
-    gasPrices.average.txFee.native = getNativeGasPrice(prices, gasPrices.average.txFee.value.amount);
-    gasPrices.slow.txFee.native = getNativeGasPrice(prices, gasPrices.slow.txFee.value.amount);
+  if (prices) {
+    gasPrices.fast.txFee.native = getNativeGasPrice(prices, gasPrices.fast.txFee.value.amount, nativeCurrency);
+    gasPrices.average.txFee.native = getNativeGasPrice(prices, gasPrices.average.txFee.value.amount, nativeCurrency);
+    gasPrices.slow.txFee.native = getNativeGasPrice(prices, gasPrices.slow.txFee.value.amount, nativeCurrency);
   }
   return nativeGases;
 };
 
-export const getNativeGasPrice = (prices, feeAmount) => {
+export const getNativeGasPrice = (prices, feeAmount, nativeCurrency) => {
+  const selected = nativeCurrencies[nativeCurrency];
   const amount = convertAssetAmountToNativeAmount(
     feeAmount,
     { symbol: 'ETH' },
     prices,
+    nativeCurrency,
   );
   return {
-    selected: prices.selected,
+    selected,
     value: {
       amount,
       display: convertAmountToDisplay(
         amount,
         prices,
         null,
+        nativeCurrency,
         2,
       ),
     },
@@ -174,26 +177,24 @@ export const getNativeGasPrice = (prices, feeAmount) => {
  * @param {Object} prices
  * @param {Number} gasLimit
  */
-export const parseGasPricesTxFee = (gasPrices, prices, gasLimit) => {
+export const parseGasPricesTxFee = (gasPrices, prices, gasLimit, nativeCurrency) => {
   gasPrices.fast.txFee = getTxFee(gasPrices.fast.value.amount, gasLimit);
   gasPrices.average.txFee = getTxFee(gasPrices.average.value.amount, gasLimit);
   gasPrices.slow.txFee = getTxFee(gasPrices.slow.value.amount, gasLimit);
-  return convertGasPricesToNative(prices, gasPrices);
+  return convertGasPricesToNative(prices, gasPrices, nativeCurrency);
 };
 
 /**
  * @desc parse prices object from api response
  * @param  {Object} [data=null]
  * @param  {Array} [assets=[]]
- * @param  {String} [native='USD']
  * @return {Object}
  */
 export const parsePricesObject = (
   data = null,
   assets = [],
-  nativeSelected = 'USD',
 ) => {
-  let prices = { selected: nativeCurrencies[nativeSelected] };
+  let prices = {};
   Object.keys(nativeCurrencies).forEach(nativeCurrency => {
     prices[nativeCurrency] = {};
     assets.forEach(asset => {
@@ -237,7 +238,7 @@ export const parsePricesObject = (
  * @desc parse account assets
  * @param  {Object} [data=null]
  * @param  {String} [address=null]
- * @return {Object}
+ * @return {Array}
  */
 export const parseAccountAssets = (data = null, address = '') => {
   try {
@@ -253,6 +254,7 @@ export const parseAccountAssets = (data = null, address = '') => {
         address: assetData.contract.address || null,
         decimals: convertStringToNumber(assetData.contract.decimals),
       };
+      // TODO: can only store balance without display
       const assetBalance = convertAssetAmountToBigNumber(
         assetData.balance,
         asset.decimals,
@@ -266,7 +268,6 @@ export const parseAccountAssets = (data = null, address = '') => {
             decimals: asset.decimals,
           }),
         },
-        native: null,
       };
     });
 
@@ -274,99 +275,10 @@ export const parseAccountAssets = (data = null, address = '') => {
       asset => !!Number(asset.balance.amount) || asset.symbol === 'ETH',
     );
 
-    return {
-      address: address,
-      type: '',
-      assets: assets,
-      total: null,
-    };
+    return assets;
   } catch (error) {
     throw error;
   }
-};
-
-/**
- * @desc parse account balances from native prices
- * @param  {Object} [account=null]
- * @param  {Object} [prices=null]
- * @param  {String} [network='']
- * @return {Object}
- */
-export const parseAccountBalancesPrices = (
-  account = null,
-  nativePrices = null,
-  network = '',
-) => {
-  let totalAmount = 0;
-  let newAccount = {
-    ...account,
-  };
-  let nativeSelected = nativePrices.selected.currency;
-  if (account) {
-    const newAssets = account.assets.map(asset => {
-      if (
-        !nativePrices ||
-        (nativePrices && !nativePrices[nativeSelected][asset.symbol])
-      )
-        return asset;
-
-      const balanceAmountUnit = convertAmountFromBigNumber(
-        asset.balance.amount,
-        asset.decimals,
-      );
-      const balancePriceUnit = convertAmountFromBigNumber(
-        nativePrices[nativeSelected][asset.symbol].price.amount,
-      );
-      const balanceRaw = multiply(balanceAmountUnit, balancePriceUnit);
-      const balanceAmount = convertAmountToBigNumber(balanceRaw);
-      let trackingAmount = balanceAmount;
-      if (nativeSelected !== 'USD') {
-        const trackingPriceUnit = convertAmountFromBigNumber(
-          nativePrices['USD'][asset.symbol].price.amount,
-        );
-        const trackingRaw = multiply(balanceAmountUnit, trackingPriceUnit);
-        trackingAmount = convertAmountToBigNumber(trackingRaw);
-      }
-      const balanceDisplay = convertAmountToDisplay(
-        balanceAmount,
-        nativePrices,
-      );
-      const assetPrice = nativePrices[nativeSelected][asset.symbol].price;
-      return {
-        ...asset,
-        trackingAmount,
-        native: {
-          selected: nativePrices.selected,
-          balance: { amount: balanceAmount, display: balanceDisplay },
-          price: assetPrice,
-          change:
-            asset.symbol === nativePrices.selected.currency
-              ? { amount: '0', display: '———' }
-              : nativePrices[nativeSelected][asset.symbol].change,
-        },
-      };
-    });
-    totalAmount = newAssets.reduce(
-      (total, asset) =>
-        add(total, asset.native ? asset.native.balance.amount : 0),
-      0,
-    );
-    const totalUSDAmount = (nativeSelected === 'USD') ? totalAmount :
-      newAssets.reduce(
-        (total, asset) =>
-          add(total, asset.native ? asset.trackingAmount : 0),
-        0,
-      );
-    const totalDisplay = convertAmountToDisplay(totalAmount, nativePrices);
-    const totalTrackingAmount = convertAmountToUnformattedDisplay(totalUSDAmount, 'USD');
-    const total = { amount: totalAmount, display: totalDisplay, totalTrackingAmount };
-    newAccount = {
-      ...newAccount,
-      assets: newAssets,
-      total: total,
-    };
-  }
-  return newAccount;
 };
 
 /**
@@ -437,7 +349,7 @@ export const parseHistoricalNativePrice = async transaction => {
   const feeResponse = historicalPriceResponses[1];
 
   Object.keys(nativeCurrencies).forEach(nativeCurrency => {
-    let prices = { selected: nativeCurrencies[nativeCurrency] };
+    let prices = {};
     prices[nativeCurrency] = {};
     if (response.data.response !== 'Error' && response.data[asset.symbol]) {
       const assetPriceAmount = convertAmountToBigNumber(
@@ -449,6 +361,8 @@ export const parseHistoricalNativePrice = async transaction => {
       const assetPriceDisplay = convertAmountToDisplay(
         assetPriceAmount,
         prices,
+        null,
+        nativeCurrency,
       );
       prices[nativeCurrency][asset.symbol].price.display = assetPriceDisplay;
       const assetPrice = prices[nativeCurrency][asset.symbol].price;
@@ -457,10 +371,13 @@ export const parseHistoricalNativePrice = async transaction => {
         tx.value.amount,
         asset,
         prices,
+        nativeCurrency,
       );
       const valuePriceDisplay = convertAmountToDisplay(
         valuePriceAmount,
         prices,
+        null,
+        nativeCurrency,
       );
       const valuePrice = !tx.error
         ? { amount: valuePriceAmount, display: valuePriceDisplay }
@@ -482,17 +399,20 @@ export const parseHistoricalNativePrice = async transaction => {
       prices[nativeCurrency]['ETH'] = {
         price: { amount: feePriceAmount, display: null },
       };
-      const feePriceDisplay = convertAmountToDisplay(feePriceAmount, prices);
+      const feePriceDisplay = convertAmountToDisplay(feePriceAmount, prices, null, nativeCurrency);
       prices[nativeCurrency]['ETH'].price.display = feePriceDisplay;
 
       const txFeePriceAmount = convertAssetAmountToNativeValue(
         tx.txFee.amount,
         ethFeeAsset,
         prices,
+        nativeCurrency,
       );
       const txFeePriceDisplay = convertAmountToDisplay(
         txFeePriceAmount,
         prices,
+        null,
+        nativeCurrency,
       );
       const txFeePrice = {
         amount: txFeePriceAmount,
@@ -551,10 +471,10 @@ export const parseNewTransaction = async (
     from: txDetails.from,
     to: txDetails.to,
     error: false,
+    native: { selected: nativeCurrencies[nativeCurrency] },
     nonce: nonce,
     value: value,
     txFee: txFee,
-    native: { selected: nativeCurrencies[nativeCurrency] },
     pending: txDetails.hash ? true : false,
     asset: txDetails.asset,
   };