+++ src/reducers/index.js
@@ -1,17 +1,29 @@
-import account from './_account';
+import assets from './_assets';
+import prices from './_prices';
 import send from './_send';
+import settings from './_settings';
+import transactions from './_transactions';
 import {
-  accountChangeLanguage,
-  accountChangeNativeCurrency,
-  accountClearState,
-  accountInitializeState,
-  accountUpdateAccountAddress,
-  accountUpdateExchange,
-  accountUpdateHasPendingTransaction,
-  accountUpdateNetwork,
-  accountUpdateTransactions,
-  INITIAL_ACCOUNT_STATE,
-} from './_account';
+  assetsClearState,
+  assetsRefreshState,
+  INITIAL_ASSETS_STATE,
+} from './_assets';
+import {
+  pricesClearState,
+} from './_prices';
+import {
+  settingsChangeLanguage,
+  settingsChangeNativeCurrency,
+  settingsInitializeState,
+  settingsUpdateAccountAddress,
+  settingsUpdateNetwork,
+} from './_settings';
+import {
+  transactionsClearState,
+  transactionsRefreshState,
+  transactionsUpdateHasPendingTransaction,
+  transactionsAddNewTransaction,
+} from './_transactions';
 import {
   sendClearFields,
   sendMaxBalance,
@@ -25,17 +37,12 @@ import {
   sendUpdateSelected,
 } from './_send';
 export {
-  account,
-  accountChangeLanguage,
-  accountChangeNativeCurrency,
-  accountClearState,
-  accountInitializeState,
-  accountUpdateAccountAddress,
-  accountUpdateExchange,
-  accountUpdateHasPendingTransaction,
-  accountUpdateNetwork,
-  accountUpdateTransactions,
-  INITIAL_ACCOUNT_STATE,
+  assets,
+  assetsClearState,
+  assetsRefreshState,
+  INITIAL_ASSETS_STATE,
+  prices,
+  pricesClearState,
   send,
   sendClearFields,
   sendMaxBalance,
@@ -47,4 +54,15 @@ export {
   sendUpdateNativeAmount,
   sendUpdateRecipient,
   sendUpdateSelected,
+  settings,
+  settingsChangeLanguage,
+  settingsChangeNativeCurrency,
+  settingsInitializeState,
+  settingsUpdateAccountAddress,
+  settingsUpdateNetwork,
+  transactions,
+  transactionsClearState,
+  transactionsRefreshState,
+  transactionsUpdateHasPendingTransaction,
+  transactionsAddNewTransaction,
 };